I hope final

+++ brevnov.nikita/P2/brevnov.hpp
@@ -5,6 +5,6 @@ namespace brevnov
 {
   double f(const double x, const size_t k, const  double error);
   void outputing(const double left, const double right, const size_t k);
-  void tablecheck(size_t i, const size_t k, const double error);
+  void tablecheck(double i, const size_t k, const double error);
 } 
 #endif
